+++ sonoff/xdrv_05_irremote.ino
@@ -175,7 +175,7 @@ void IrReceiveCheck(void)
           if (strlen(mqtt_data) > sizeof(mqtt_data) - 40) { break; }  // Quit if char string becomes too long
         }
         uint16_t extended_length = results.rawlen - 1;
-        for (uint16_t j = 0; j < results.rawlen - 1; j++) {
+        for (uint32_t j = 0; j < results.rawlen - 1; j++) {
           uint32_t usecs = results.rawbuf[j] * kRawTick;
           // Add two extra entries for multiple larger than UINT16_MAX it is.
           extended_length += (usecs / (UINT16_MAX + 1)) * 2;
@@ -261,7 +261,7 @@ uint8_t IrHvacToshiba(const char *HVAC_Mode, const char *HVAC_FanMode, bool HVAC
   data[5] = (uint8_t)(Temp - 17) << 4;
 
   data[HVAC_TOSHIBA_DATALEN - 1] = 0;
-  for (int x = 0; x < HVAC_TOSHIBA_DATALEN - 1; x++) {
+  for (uint32_t x = 0; x < HVAC_TOSHIBA_DATALEN - 1; x++) {
     data[HVAC_TOSHIBA_DATALEN - 1] = (uint8_t)data[x] ^ data[HVAC_TOSHIBA_DATALEN - 1]; // CRC is a simple bits addition
   }
 
@@ -273,7 +273,7 @@ uint8_t IrHvacToshiba(const char *HVAC_Mode, const char *HVAC_FanMode, bool HVAC
   rawdata[i++] = HVAC_TOSHIBA_HDR_SPACE;
 
   //data
-  for (int b = 0; b < HVAC_TOSHIBA_DATALEN; b++) {
+  for (uint32_t b = 0; b < HVAC_TOSHIBA_DATALEN; b++) {
     for (mask = B10000000; mask > 0; mask >>= 1) { //iterate through bit mask
       if (data[b] & mask) { // Bit ONE
         rawdata[i++] = HVAC_TOSHIBA_BIT_MARK;
@@ -447,7 +447,7 @@ uint8_t IrHvacLG(const char *HVAC_Mode, const char *HVAC_FanMode, bool HVAC_Powe
   }
   // Build LG IR code
   LG_Code = data[0] << 4;
-  for (int i = 1; i < 6; i++) {
+  for (uint32_t i = 1; i < 6; i++) {
     LG_Code = (LG_Code + data[i]) << 4;
   }
   LG_Code = LG_Code + data[6];
@@ -568,7 +568,7 @@ bool IrSendCommand(void)
               error = IE_INVALID_RAWDATA;
             } else {
               uint16_t parm[count];
-              for (uint8_t i = 0; i < count; i++) {
+              for (uint32_t i = 0; i < count; i++) {
                 parm[i] = strtol(strtok_r(nullptr, ", ", &p), nullptr, 0);
                 if (!parm[i]) {
                   if (!i) {
@@ -641,7 +641,7 @@ bool IrSendCommand(void)
               count++;
               if (count < 200) {
                 uint16_t raw_array[count];  // It's safe to use stack for up to 200 packets (limited by mqtt_data length)
-                for (uint16_t i = 0; i < count; i++) {
+                for (uint32_t i = 0; i < count; i++) {
                   raw_array[i] = strtol(strtok_r(nullptr, ", ", &p), nullptr, 0);  // Allow decimal (20496) and hexadecimal (0x5010) input
                 }
 
@@ -654,7 +654,7 @@ bool IrSendCommand(void)
                 if (raw_array == nullptr) {
                   error = IE_INVALID_RAWDATA;
                 } else {
-                  for (uint16_t i = 0; i < count; i++) {
+                  for (uint32_t i = 0; i < count; i++) {
                     raw_array[i] = strtol(strtok_r(nullptr, ", ", &p), nullptr, 0);  // Allow decimal (20496) and hexadecimal (0x5010) input
                   }
